+++ gaios/Helpers/Constants.swift
@@ -9,4 +9,5 @@ enum Constants {
 enum AppStorage {
     static let dontShowTorAlert = "dont_show_tor_alert"
     static let defaultTransactionPriority = "default_transaction_priority"
+    static let testnetIsVisible = "testnet_is_visible"
 }