+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Reset two factor authentication
 - Refactoring restore wallet flow, enabled Liquid Testnet in production, UI fix for iOS15
+- App Settings: new option to show/hide testnet selection during onboarding
 
 ### Changed